invoice club logo,name,address fixing, All modals image fixing, locker ocuupied fixing.

+++ app/Http/Controllers/Master/LockerManageController.php
@@ -144,6 +144,10 @@ class LockerManageController extends Controller
                          ->where('id', $id)
                          ->firstOrFail();
 
+        if( $lockers->status == 'occupied'){
+            return redirect()->route('manage-lockers.index')->with('error', 'This locker is already assigned to a member and cannot be deleted.');
+        }
+
         $lockers->delete();
 
         return redirect()
